requirements.txt
Pillow==10.0.0
Pillow==10.3.0
pytube==15.0.0
Requests==2.32.3
streamlit==1.35.0
